add simultaneous option to README and usage
Add -s --simultaneous XYZ option to readme and
usage. Defaults to Infinity.

It restricts the number of connections to a
specific resource. So if you set it to 1, then
it will only do one connection at a time to
each resource, waiting for the connection to
succeed of fail before making next one.

Update the text about stability window, since
it only applies to file sizes in the new
version.