force axios to use http adapter
If jest/jsdom is used then a polyfill for xhr is created which will cause axios to default the adapter to it. Since jsdom logs all errors use of wait-on will have any failed attempts to connect logged which is rather noisy and misleading.

By forcing axios to use the node.js http adapter instead this problem is bypassed keeping logs clean.